requirements.txt
flake8
pydicom
pytest
python-dotenv
requests
